75: moved helper in a separate package
Task-Url: http://github.com/LorenzoBettini/edelta/issues/75
