+++ edelta.parent/edelta.refactorings.lib/META-INF/MANIFEST.MF
@@ -9,5 +9,6 @@ Require-Bundle: edelta.lib;visibility:=reexport
 Import-Package: org.apache.log4j,
  org.apache.commons.logging
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
-Export-Package: edelta.refactorings.lib
+Export-Package: edelta.refactorings.lib,
+ edelta.refactorings.lib.helper
 Automatic-Module-Name: edelta.refactorings.lib
